// TEST_PROJECT/API/dataobject/object.types.ts
export interface User {
  firstName: string
  lastName: string
  email: string
  password: string
}

export interface Contact {
  firstName: string
  lastName: string
  birthdate: string
  email: string
  phone: string
  street1: string
  street2: string
  city: string
  stateProvince: string
  postalCode: string
  country: string
}